refactor(roles): extract name glob filter into shared v2_filters module
Move _glob_to_regex and the name filtering logic from RoleV2QuerySet into
a new management.v2_filters module as v2_name_filter(). This shared utility
supports case-insensitive exact match (no wildcards), glob pattern matching
(with * wildcards), and a bare * that matches everything.

RoleV2QuerySet.named() now delegates to v2_name_filter() with identical
behavior. The module is kept dependency-free (only django.db.models) to
avoid circular imports through management.models.
